chore(api): move solution-service-qroc.js to evaluation scope
